split db implementation and other stuff
